Don't handle white points in RGB<->XYZ conversions
An RGB space with white point A should not have an RGB space with white
point B as its base. Instead, spaces and their bases should always have
the same white point, which is ultimately determined by the first XYZ
space in the chain. For example, sRGB leads to XYZ D65, and ProPhoto
leads to XYZ D50. A theoretical ProPhoto D65 should not have ProPhoto as
its base.